chore: remove obsolete routes file
