+++ server/__tests__/characters.test.js
@@ -3,11 +3,17 @@ const express = require('express');
 
 jest.mock('../db/conn');
 const dbo = require('../db/conn');
-const charactersRouter = require('../routes.js');
+const campaignsRouter = require('../routes/campaigns');
+const charactersRouter = require('../routes/characters');
 
 const app = express();
 app.use(express.json());
-app.use(charactersRouter);
+app.use(async (req, res, next) => {
+  req.db = await dbo();
+  next();
+});
+app.use('/campaigns', campaignsRouter);
+app.use('/characters', charactersRouter);
 
 describe('Character routes', () => {
   test('add character success', async () => {
@@ -17,7 +23,7 @@ describe('Character routes', () => {
       })
     });
     const res = await request(app)
-      .post('/character/add')
+      .post('/characters/add')
       .send({ token: 'alice', characterName: 'Hero', campaign: 'Camp1' });
     expect(res.status).toBe(200);
     expect(res.body.acknowledged).toBe(true);
@@ -45,7 +51,7 @@ describe('Character routes', () => {
       newSkill: ['Stealth']
     };
     const res = await request(app)
-      .post('/character/add')
+      .post('/characters/add')
       .send(payload);
     expect(res.status).toBe(200);
     expect(captured).toMatchObject({
@@ -63,7 +69,7 @@ describe('Character routes', () => {
       })
     });
     const res = await request(app)
-      .post('/character/add')
+      .post('/characters/add')
       .send({ token: 'alice', characterName: 'Hero', campaign: 'Camp1' });
     expect(res.status).toBe(500);
   });
@@ -75,7 +81,7 @@ describe('Character routes', () => {
       })
     });
     const res = await request(app)
-      .post('/character/add')
+      .post('/characters/add')
       .send({ token: 'alice' });
     expect(res.status).toBe(400);
   });
@@ -86,7 +92,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(null, [{ token: 'alice', campaign: 'Camp1' }]) })
       })
     });
-    const res = await request(app).get('/campaign/Camp1/alice');
+    const res = await request(app).get('/campaigns/Camp1/alice');
     expect(res.status).toBe(200);
     expect(res.body[0].token).toBe('alice');
   });
@@ -97,7 +103,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(new Error('db error')) })
       })
     });
-    const res = await request(app).get('/campaign/Camp1/alice');
+    const res = await request(app).get('/campaigns/Camp1/alice');
     expect(res.status).toBe(500);
   });
 
@@ -110,7 +116,7 @@ describe('Character routes', () => {
         ]) })
       })
     });
-    const res = await request(app).get('/campaign/Camp1/characters');
+    const res = await request(app).get('/campaigns/Camp1/characters');
     expect(res.status).toBe(200);
     expect(res.body.length).toBe(2);
   });
@@ -121,7 +127,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(new Error('db error')) })
       })
     });
-    const res = await request(app).get('/campaign/Camp1/characters');
+    const res = await request(app).get('/campaigns/Camp1/characters');
     expect(res.status).toBe(500);
   });
 
@@ -154,7 +160,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(null, [{ weaponName: 'Sword' }]) })
       })
     });
-    const res = await request(app).get('/weapons/Camp1');
+    const res = await request(app).get('/characters/weapons/Camp1');
     expect(res.status).toBe(200);
     expect(res.body[0].weaponName).toBe('Sword');
   });
@@ -165,7 +171,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(new Error('db error')) })
       })
     });
-    const res = await request(app).get('/weapons/Camp1');
+    const res = await request(app).get('/characters/weapons/Camp1');
     expect(res.status).toBe(500);
   });
 
@@ -175,7 +181,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(null, [{ armorName: 'Plate' }]) })
       })
     });
-    const res = await request(app).get('/armor/Camp1');
+    const res = await request(app).get('/characters/armor/Camp1');
     expect(res.status).toBe(200);
     expect(res.body[0].armorName).toBe('Plate');
   });
@@ -186,7 +192,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(new Error('db error')) })
       })
     });
-    const res = await request(app).get('/armor/Camp1');
+    const res = await request(app).get('/characters/armor/Camp1');
     expect(res.status).toBe(500);
   });
 
@@ -196,7 +202,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(null, [{ itemName: 'Potion' }]) })
       })
     });
-    const res = await request(app).get('/items/Camp1');
+    const res = await request(app).get('/characters/items/Camp1');
     expect(res.status).toBe(200);
     expect(res.body[0].itemName).toBe('Potion');
   });
@@ -207,99 +213,7 @@ describe('Character routes', () => {
         find: () => ({ toArray: (cb) => cb(new Error('db error')) })
       })
     });
-    const res = await request(app).get('/items/Camp1');
-    expect(res.status).toBe(500);
-  });
-
-  test('get feats success', async () => {
-    dbo.mockResolvedValue({
-      collection: () => ({
-        find: () => ({ toArray: (cb) => cb(null, [{ feat: 'Power Attack' }]) })
-      })
-    });
-    const res = await request(app).get('/feats');
-    expect(res.status).toBe(200);
-    expect(res.body[0].feat).toBe('Power Attack');
-  });
-
-  test('get feats failure', async () => {
-    dbo.mockResolvedValue({
-      collection: () => ({
-        find: () => ({ toArray: (cb) => cb(new Error('db error')) })
-      })
-    });
-    const res = await request(app).get('/feats');
-    expect(res.status).toBe(500);
-  });
-
-  test('get occupations success', async () => {
-    dbo.mockResolvedValue({
-      collection: () => ({
-        find: () => ({ toArray: (cb) => cb(null, [{ name: 'Soldier' }]) })
-      })
-    });
-    const res = await request(app).get('/occupations');
-    expect(res.status).toBe(200);
-    expect(res.body[0].name).toBe('Soldier');
-  });
-
-  test('get occupations failure', async () => {
-    dbo.mockResolvedValue({
-      collection: () => ({
-        find: () => ({ toArray: (cb) => cb(new Error('db error')) })
-      })
-    });
-    const res = await request(app).get('/occupations');
-    expect(res.status).toBe(500);
-  });
-
-  test('update skills success', async () => {
-    dbo.mockResolvedValue({
-      collection: () => ({
-        findOneAndUpdate: async () => ({ value: { appraise: 1 } })
-      })
-    });
-    const res = await request(app)
-      .put('/update-skills/507f1f77bcf86cd799439011')
-      .send({ appraise: 1 });
-    expect(res.status).toBe(200);
-    expect(res.body.appraise).toBe(1);
-  });
-
-  test('update skills failure', async () => {
-    dbo.mockResolvedValue({
-      collection: () => ({
-        findOneAndUpdate: async () => { throw new Error('db error'); }
-      })
-    });
-    const res = await request(app)
-      .put('/update-skills/507f1f77bcf86cd799439011')
-      .send({ appraise: 1 });
-    expect(res.status).toBe(500);
-  });
-
-  test('update added skills success', async () => {
-    dbo.mockResolvedValue({
-      collection: () => ({
-        findOneAndUpdate: async () => ({ value: { newSkill: [['Skill', 1]] } })
-      })
-    });
-    const res = await request(app)
-      .put('/updated-add-skills/507f1f77bcf86cd799439011')
-      .send({ newSkill: [['Skill', 1]] });
-    expect(res.status).toBe(200);
-    expect(res.body.newSkill).toEqual([['Skill', 1]]);
-  });
-
-  test('update added skills failure', async () => {
-    dbo.mockResolvedValue({
-      collection: () => ({
-        findOneAndUpdate: async () => { throw new Error('db error'); }
-      })
-    });
-    const res = await request(app)
-      .put('/updated-add-skills/507f1f77bcf86cd799439011')
-      .send({ newSkill: [['Skill', 1]] });
+    const res = await request(app).get('/characters/items/Camp1');
     expect(res.status).toBe(500);
   });
 
@@ -308,7 +222,7 @@ describe('Character routes', () => {
       collection: () => ({ insertOne: (doc, cb) => cb(null, { acknowledged: true }) })
     });
     const res = await request(app)
-      .post('/weapon/add')
+      .post('/characters/weapon/add')
       .send({ campaign: 'Camp1', weaponName: 'Sword' });
     expect(res.status).toBe(200);
     expect(res.body.acknowledged).toBe(true);
@@ -319,7 +233,7 @@ describe('Character routes', () => {
       collection: () => ({ insertOne: (doc, cb) => cb(new Error('db error')) })
     });
     const res = await request(app)
-      .post('/weapon/add')
+      .post('/characters/weapon/add')
       .send({ campaign: 'Camp1', weaponName: 'Sword' });
     expect(res.status).toBe(500);
   });
@@ -329,7 +243,7 @@ describe('Character routes', () => {
       collection: () => ({ insertOne: (doc, cb) => cb(null, { acknowledged: true }) })
     });
     const res = await request(app)
-      .post('/armor/add')
+      .post('/characters/armor/add')
       .send({ campaign: 'Camp1', armorName: 'Plate' });
     expect(res.status).toBe(200);
     expect(res.body.acknowledged).toBe(true);
@@ -340,7 +254,7 @@ describe('Character routes', () => {
       collection: () => ({ insertOne: (doc, cb) => cb(new Error('db error')) })
     });
     const res = await request(app)
-      .post('/armor/add')
+      .post('/characters/armor/add')
       .send({ campaign: 'Camp1', armorName: 'Plate' });
     expect(res.status).toBe(500);
   });
@@ -350,7 +264,7 @@ describe('Character routes', () => {
       collection: () => ({ insertOne: (doc, cb) => cb(null, { acknowledged: true }) })
     });
     const res = await request(app)
-      .post('/item/add')
+      .post('/characters/item/add')
       .send({ campaign: 'Camp1', itemName: 'Potion' });
     expect(res.status).toBe(200);
     expect(res.body.acknowledged).toBe(true);
@@ -361,7 +275,7 @@ describe('Character routes', () => {
       collection: () => ({ insertOne: (doc, cb) => cb(new Error('db error')) })
     });
     const res = await request(app)
-      .post('/item/add')
+      .post('/characters/item/add')
       .send({ campaign: 'Camp1', itemName: 'Potion' });
     expect(res.status).toBe(500);
   });
